Connection between Event Flow and Speaker Detail Content
