+++ lib/event_flow/ui/session_info_field.dart
@@ -1,9 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+
 import 'package:festival_flutterturkiye_org/core/model/session.dart';
 import 'package:festival_flutterturkiye_org/core/model/speaker.dart';
+import 'package:festival_flutterturkiye_org/core/ui/speaker_image.dart';
+import 'package:festival_flutterturkiye_org/core/utils/dialog_helper.dart';
 import 'package:festival_flutterturkiye_org/core/utils/theme_helper.dart';
 import 'package:festival_flutterturkiye_org/event_flow/ui/session_time_field.dart';
-import 'package:flutter/material.dart';
-import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 
 class SessionInfoField extends StatelessWidget {
   const SessionInfoField({
@@ -92,60 +96,29 @@ class _EventFlowSpeaker extends StatelessWidget {
     if (speaker == null) {
       return const SizedBox.shrink();
     }
-    return GestureDetector(
-      onTap: () {
-        // TODO: It will connect to Speaker Detail Issue: #57
-        debugPrint('Go To Speaker Detail');
-      },
-      child: Row(
-        children: [
-          _SpeakerImage(
-            speakerImage: speaker.image,
-            imageSize: 64,
-          ),
-          const SizedBox(width: 8),
-          Text(
-            speaker.name,
-            style: const TextStyle(
-              color: ThemeHelper.speakerDetailImageBorder,
-              fontSize: 16,
-              fontWeight: FontWeight.w500,
+    return MouseRegion(
+      cursor: SystemMouseCursors.click,
+      child: GestureDetector(
+        onTap: () => DialogHelper.showSpeaker(context, speaker: speaker),
+        child: Row(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            SpeakerImage(
+              speakerImage: speaker.image,
+              imageSize: 64,
             ),
-          )
-        ],
+            const SizedBox(width: 8),
+            Text(
+              speaker.name,
+              style: const TextStyle(
+                color: ThemeHelper.speakerDetailImageBorder,
+                fontSize: 16,
+                fontWeight: FontWeight.w500,
+              ),
+            )
+          ],
+        ),
       ),
     );
   }
-}
-
-class _SpeakerImage extends StatelessWidget {
-  const _SpeakerImage({
-    @required this.speakerImage,
-    @required this.imageSize,
-    Key key,
-  })  : assert(speakerImage != null),
-        assert(imageSize != null),
-        super(key: key);
-
-  final String speakerImage;
-  final double imageSize;
-
-  @override
-  Widget build(BuildContext context) => SizedBox(
-        height: imageSize,
-        width: imageSize,
-        child: DecoratedBox(
-          decoration: BoxDecoration(
-            image: DecorationImage(
-              image: NetworkImage(speakerImage),
-              fit: BoxFit.cover,
-            ),
-            shape: BoxShape.circle,
-            border: Border.all(
-              color: ThemeHelper.speakerDetailImageBorder,
-              width: 2,
-            ),
-          ),
-        ),
-      );
 }
